bootloader: use _PyObject_CallFunction_SizeT on Python 3.10
When running under python 3.10 or later, we need to map the
PI_PyObject_CallFunction to _PyObject_CallFunction_SizeT
instead of PyObject_CallFunction, as python 3.10 requires
the 'y#' format to use Py_ssize_t-based string length
(see PEP 353).

In general, we need to take greater care on how we pass
buffer sizes to PI_PyObject_CallFunction, as this is done
via variadic argument list, and a mismatch in the argument
type size (after default argument promotion) on our side
and the expected argument type size on python's side can
result in undefined behavior.

Therefore, for python 3.9 and earlier, we explicitly pass
size as (32-bit) int type (as PyObject_CallFunction expects
int argument) - in case we ever change ulen field in TOC to
64-bit... And for python 3.10 and later, we pass the size as
size_t (which, same as ssize_t, matches the size of Py_ssize_t;
32-bit on 32-bit platforms, and 64-bit on 64-bit platforms).