MoreSpeed!: Use faster importlib method to load modules.
Using this implemention, as I understand it, allows importlib to
internally maintain a cached index of os.listdir(path) for each path in
sys.path so that it rarely has to rescan a folder.
The old implementation forced it to rescan every path in sys.path and
convert/filter every filename to module name each time we load 1 module.

This knocks a good 20% off the build time by my benchmarks.